+++ implementation/site/theme-mockups.js
@@ -1,15 +1,31 @@
 (function(app){
 
 	app.context('_Mockups', {
+		
+		className: 'wrapper-full container-fluid',
 
 		mockups: [
-			'nav-bar.html'
+			//navbars
+			{tpl: 'nav-bar.html', className: 'navbar-default'},
+			{tpl: 'nav-bar.html', className: 'navbar-inverse'},
+
+			//headings
+			{tpl: 'heading.html', className: 'heading'},
+
+			//boxes
+			{tpl: 'boxes.html', className:'row'}
+
 		],
+
 		onShow: function(){
-			_.each(this.mockups, function(name){
+			_.each(this.mockups, function(m){
 
 				this.$el.append(app.view({
-					template: '@mockups/' + name
+					className: 'wrapper-full',
+					template: '@mockups/' + m.tpl,
+					onRender: function(){
+						this.$el.find('> div').addClass(m.className);
+					}
 				}, true).render().el);
 
 			}, this);