+++ CHANGELOG.md
@@ -3,8 +3,10 @@ Change Log
 1.6.0 (2014-07-24*)
 -------------------
 1. Removed default actions settings from datagrid action cell;
-2. Added wrapper style;
+2. Added wrapper style (full, x2);
 3. Added btn-outline style;
+4. Added heading style;
+5. Added box style (heading, body);
 
 In progress...